// src/generated/models/Dnx_assessment_versionsModel.ts
/*!
 * Copyright (C) Microsoft Corporation. All rights reserved.
 * This file is autogenerated. Do not edit this file directly.
 */
export const Dnx_assessment_versionsstatecode = {
  0: 'Active',
  1: 'Inactive'
} as const;
export type Dnx_assessment_versionsstatecode = keyof typeof Dnx_assessment_versionsstatecode;
export const Dnx_assessment_versionsstatuscode = {
  1: 'Active',
  2: 'Inactive'
} as const;
export type Dnx_assessment_versionsstatuscode = keyof typeof Dnx_assessment_versionsstatuscode;

export interface Dnx_assessment_versionsBase {
  "dnx_Assessment@odata.bind"?: string;
  dnx_assessment_versionid: string;
  dnx_change_summary?: string;
  dnx_version_number: string;
  importsequencenumber?: number;
  overriddencreatedon?: string;
  ownerid: string;
  owneridtype: string;
  statecode: Dnx_assessment_versionsstatecode;
  statuscode?: Dnx_assessment_versionsstatuscode;
  timezoneruleversionnumber?: number;
  utcconversiontimezonecode?: number;
}

export interface Dnx_assessment_versions extends Dnx_assessment_versionsBase {
  createdbyname?: string;
  createdbyyominame: string;
  createdon?: string;
  createdonbehalfbyname?: string;
  createdonbehalfbyyominame: string;
  dnx_assessmentname?: string;
  dnx_snapsho_tjson?: string;
  dnx_snapsho_tjson_name?: string;
  modifiedbyname?: string;
  modifiedbyyominame: string;
  modifiedon?: string;
  modifiedonbehalfbyname?: string;
  modifiedonbehalfbyyominame: string;
  owneridname: string;
  owneridyominame: string;
  owningbusinessunitname: string;
  statecodename?: string;
  statuscodename?: string;
  versionnumber?: number;
  createdby?: object;
  _createdby_value?: string;
  createdonbehalfby?: object;
  _createdonbehalfby_value?: string;
  dnx_assessment?: object;
  _dnx_assessment_value?: string;
  modifiedby?: object;
  _modifiedby_value?: string;
  modifiedonbehalfby?: object;
  _modifiedonbehalfby_value?: string;
  owningbusinessunit?: object;
  _owningbusinessunit_value?: string;
  owningteam?: object;
  _owningteam_value?: string;
  owninguser?: object;
  _owninguser_value?: string;
}
export type Dnx_assessment_versionsUploadColumnName = 'dnx_snapsho_tjson';